packages/devhub-core/pyproject.toml
[project]
name = "devhub-core"
version = "0.2.0"
description = "Core contracts, adapters and agents for the DevHub orchestration platform"
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
